requirements.txt
requests
python-dotenv
fastapi==0.115.14
uvicorn
SQLAlchemy==2.0.41
psycopg2==2.9.10